登录功能、jsr303参数校验

+++ src/main/java/com/imooc/miaosha/dao/UserDao.java
@@ -10,7 +10,7 @@ import com.imooc.miaosha.domain.User;
 @Mapper
 public interface UserDao {
 	@Select("select * from user where id = #{id}")
-	public User getById(@Param("id")int id);
+	public User getById(@Param("id")long id);
 	
 	@Insert("insert into user(id, name) values(#{id}, #{name})")
 	public int insert(User user);
